Add styles for markdown code examples (inline and code blocks)
It still feels a little bit hacky currently, as some of the changes
that made inline code look decent looked pretty ugly with the
larger code blocks.  See if there is a fancy rendering option you
haven't found yet, and if not find a way to target code elements
more cleanly.