Update manifests
This is appearing more and more like something that could work out of
the box for someone other than myself so I plan on eventually not
committing these files to this exact repo.  The goal is to have these
files be generated with a single command based on naming conventions
and metadata.